requirements.txt
streamlit
langchain-groq
langchain-google-community
python-dotenv